Commit of a transaction now leaves a clean state; next: implement recovery

+++ difs-core/src/main/java/org/aksw/jena_sparql_api/txn/FileUtilsX.java
@@ -15,18 +15,19 @@ public class FileUtilsX {
 	}
 
 	public static void deleteEmptyFolders(Path path, Path baseFolder) {
-		if (!Files.isDirectory(path)) {
-			throw new IllegalArgumentException("Path must be a directory: " + path);
-		}
-
 		while (path.startsWith(baseFolder)) {
-			try {
-				Files.deleteIfExists(path);
-			} catch (IOException e) {
-				// Ignore
-				break;
-			}
-			
+			if (Files.exists(path)) {
+				if (!Files.isDirectory(path)) {
+					throw new IllegalArgumentException("Path must be a directory: " + path);
+				}
+	
+				try {
+					Files.deleteIfExists(path);
+				} catch (IOException e) {
+					// Ignore
+					break;
+				}
+			}				
 			path = path.getParent();
 		}
 	}
